+++ src/types/types.ts
@@ -642,6 +642,11 @@ export interface EdgeTransaction {
   parentNetworkFee?: string
 }
 
+export interface EdgeTransactionEvent {
+  isNew: boolean
+  transaction: EdgeTransaction
+}
+
 export interface EdgeSpendTarget {
   nativeAmount?: string
   otherParams?: JsonObject
@@ -894,6 +899,8 @@ export interface EdgeCurrencyEngineCallbacks {
     tokenId: EdgeTokenId,
     balance: string
   ) => void
+  readonly onTransactions: (transactionEvents: EdgeTransactionEvent[]) => void
+  /** @deprecated Use onTransactions */
   readonly onTransactionsChanged: (transactions: EdgeTransaction[]) => void
   readonly onTxidsChanged: (txids: EdgeTxidMap) => void
   readonly onUnactivatedTokenIdsChanged: (unactivatedTokenIds: string[]) => void